# Commercial/laravel/admin/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Igronik-lk API",
        "version": "1.0"
    },
    "paths": {
        "/api/aboutList?sort={direction}": {
            "get": {
                "tags": [
                    "About"
                ],
                "summary": "Get full list of about blocks",
                "description": "Get full list of about blocks",
                "operationId": "getAbout",
                "parameters": [
                    {
                        "name": "direction",
                        "in": "path",
                        "description": "Direction of sort (asc/desc). Default- asc by ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return json of about blocks array"
                    }
                }
            }
        },
        "/api/aboutListPage?page={pageNum}": {
            "get": {
                "tags": [
                    "About"
                ],
                "summary": "Get a list of about blocks",
                "description": "Get a list of about blocks",
                "operationId": "getAboutByPage",
                "parameters": [
                    {
                        "name": "pageNum",
                        "in": "path",
                        "description": "Page of about blocks list that needs to be fetched",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return json of about blocks list array"
                    }
                }
            }
        },
        "/api/careersList?sort={direction}": {
            "get": {
                "tags": [
                    "Careers"
                ],
                "summary": "Get full list of career values",
                "description": "Get full list of career values",
                "operationId": "getCareers",
                "parameters": [
                    {
                        "name": "direction",
                        "in": "path",
                        "description": "Direction of sort (asc/desc). Default- asc by ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return json of career values array"
                    }
                }
            }
        },
        "/api/careersListPage?page={pageNum}": {
            "get": {
                "tags": [
                    "Careers"
                ],
                "summary": "Get a list of career values by page",
                "description": "Get a list of career values by page",
                "operationId": "getCareersByPage",
                "parameters": [
                    {
                        "name": "pageNum",
                        "in": "path",
                        "description": "Page of list values that needs to be fetched",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return array of list values on career page as json"
                    }
                }
            }
        },
...
        "/api/favoursList?sort={direction}": {
            "get": {
                "tags": [
                    "Favours"
                ],
                "summary": "Get full list of favours",
                "description": "Get full list of favours",
                "operationId": "getFavours",
                "parameters": [
                    {
                        "name": "direction",
                        "in": "path",
                        "description": "Direction of sort (asc/desc). Default- asc by ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return json of favours list array favours{$key=>[...,medias[],details[],posters[]]}. Posters[] have a structure [posterID=>posterData[]]."
                    }
                }
            }
        },
        "/api/favourByKey?key={key}": {
            "get": {
                "tags": [
                    "Favours"
                ],
                "summary": "Get favour by key",
                "description": "Get favour by key",
                "operationId": "getFavourByKey",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Key of favour that needs to be fetched",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return json of favour{...,medias[],details[],posters[]}. Posters[] have a structure [posterID=>posterData[]]"
                    }
                }
            }
        },
...
    }
}